static server: Add a simple static page server
Introduce two new options for the AuthService:
* The `HOMEPAGE_URL` defines a homepage that the user will be redirected
  to when visiting the redirect URL without any query args.
  This is because some Providers (e.g., Gitlab) prompt users to access the
  client's redirect url to take a look at who the provider is, in order to
  confirm where the auth_code will be sent.
* The `AFTER_LOGOUT_URL` defines a page where the user will be
  redirected to after a successful logout.

Add a web server with default pages for homepage and after
logout.
It is served at port 8082 by default.
The web server uses templates, whose behavior is described in the
included guide.

Rename STATIC_DESTINATION_URL to AFTER_LOGIN_REDIRECT_URL because the
old name was not descriptive.
Backwards compatibility is kept by falling back to the old name, if the
new one isn't found.

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>